refactor: split StoryChunk into StoryChunk and StoryChunkHVL to isolate HDF5 dependency
